+++ src/odhcpd.h
@@ -584,9 +584,6 @@ ssize_t dhcpv6_ia_handle_IAs(uint8_t *buf, size_t buflen, struct interface *ifac
 int dhcpv6_ia_init(void);
 int dhcpv6_ia_setup_interface(struct interface *iface, bool enable);
 void dhcpv6_free_lease(struct dhcpv6_lease *lease);
-void dhcpv6_ia_enum_addrs(struct interface *iface, struct dhcpv6_lease *lease,
-			  time_t now, dhcpv6_binding_cb_handler_t func, void *arg);
-void dhcpv6_ia_write_statefile(void);
 
 int netlink_add_netevent_handler(struct netevent_handler *hdlr);
 ssize_t netlink_get_interface_addrs(const int ifindex, bool v6,
